Refactor CameraControls component
