+++ frontend/src/layouts/cameras/components/CameraControls/props.tsx
@@ -1,0 +1,6 @@
+import { Control } from "../../types";
+
+export default interface ControlProps {
+    control: Control;
+    index: number;
+}
